+++ src/Identity/Logitar.Identity.Core/Sessions/ISessionService.cs
@@ -6,5 +6,6 @@ namespace Logitar.Identity.Core.Sessions;
 public interface ISessionService
 {
   Task<Session> CreateAsync(CreateSessionPayload payload, CancellationToken cancellationToken = default);
+  Task<Session> RenewAsync(RenewPayload payload, CancellationToken cancellationToken = default);
   Task<Session> SignInAsync(SignInPayload payload, CancellationToken cancellationToken = default);
 }