+++ src/Identity/Logitar.Identity.Core/Users/IUserQuerier.cs
@@ -8,4 +8,5 @@ public interface IUserQuerier
   Task<User> ReadAsync(UserAggregate user, CancellationToken cancellationToken = default);
   Task<User?> ReadAsync(string id, CancellationToken cancellationToken = default);
   Task<User?> ReadAsync(string? tenantId, string uniqueName, CancellationToken cancellationToken = default);
+  Task<IEnumerable<User>> ReadAsync(string? tenantId, IEmailAddress email, CancellationToken cancellationToken = default);
 }