+++ src/Identity/Logitar.Identity.EntityFrameworkCore.SqlServer/Entities/UserEntity.cs
@@ -70,7 +70,12 @@ public record UserEntity : AggregateEntity
 
   public List<SessionEntity> Sessions { get; private set; } = new();
 
-  public void Authenticate(UserAuthenticatedEvent authenticated) => AuthenticatedOn = authenticated.OccurredOn;
+  public void Authenticate(UserAuthenticatedEvent authenticated)
+  {
+    SetVersion(authenticated);
+
+    AuthenticatedOn = authenticated.OccurredOn;
+  }
 
   public void ChangePassword(UserPasswordChangedEvent change, ActorEntity actor)
   {
@@ -102,7 +107,12 @@ public record UserEntity : AggregateEntity
     IsDisabled = false;
   }
 
-  public void SignIn(UserSignedInEvent signedIn) => AuthenticatedOn = signedIn.OccurredOn;
+  public void SignIn(UserSignedInEvent signedIn)
+  {
+    SetVersion(signedIn);
+
+    AuthenticatedOn = signedIn.OccurredOn;
+  }
 
   public void Update(UserUpdatedEvent updated, ActorEntity actor)
   {