+++ CHANGELOG.md
@@ -9,9 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Identity Session creation & sign-in.
+- Identity Session management.
 - Identity Token management.
-- Identity User management & authentication.
+- Identity User management.
 - Implemented query JOINs.
 
 ### Changed